Refactor preproc to use xr.Dataset as data storage (instead of loose variables)
