+++ src/wam2layers/cli.py
@@ -22,7 +22,7 @@ from wam2layers.analysis import visualization
 from wam2layers.config import Config
 from wam2layers.download import download_from_doi
 from wam2layers.example_cases import AVAILABLE_CASES
-from wam2layers.preprocessing.era5 import prep_experiment
+from wam2layers.preprocessing.shared import prep_experiment
 from wam2layers.tracking.backtrack import run_experiment as run_backtrack_experiment
 from wam2layers.tracking.forwardtrack import (
     run_experiment as run_forwardtrack_experiment,
