+++ src/wam2layers/preprocessing/era5.py
@@ -8,16 +8,10 @@ import pandas as pd
 import xarray as xr
 
 from wam2layers import __version__
-from wam2layers.config import Config
-from wam2layers.preprocessing.shared import (
+from wam2layers.preprocessing.utils import (
     accumulation_to_flux,
-    add_bounds,
-    calculate_humidity,
-    interpolate,
-    sortby_ndarray,
+    midpoints,
 )
-from wam2layers.preprocessing.xarray_append import append_to_netcdf
-from wam2layers.reference.variables import ERA5_INPUT_DATA_ATTRIBUTES
 
 logger = logging.getLogger(__name__)
 
@@ -108,11 +102,6 @@ def load_era5_ab():
     return a, b
 
 
-def midpoints(x):
-    """Linearly interpolate between the values of an array."""
-    return (x[1:] + x[:-1]) / 2
-
-
 def get_edges(era5_modellevels):
     """Get the values of a and b at the edges of a subset of ERA5 modellevels."""
     a, b = load_era5_ab()
@@ -154,302 +143,3 @@ def get_dp_modellevels(sp, levels):
     dp = p_edge.diff("level")  # in Pa
 
     return dp
-
-
-def extend_pressurelevels(
-    q: xr.DataArray,
-    u: xr.DataArray,
-    v: xr.DataArray,
-    ps: xr.DataArray,
-    qs: xr.DataArray,
-    us: xr.DataArray,
-    vs: xr.DataArray,
-) -> tuple[xr.DataArray, xr.DataArray, xr.DataArray, xr.DataArray, xr.DataArray]:
-    """Extend the levels to top of atmosphere, surface and 2-layer boundary level.
-
-    The input data does not have data at the surface and
-    and the top of atmosphere. The input data is extended to
-    these edges.
-
-    For proper integration, extra layers have to be inserted
-    near the 2-layer boundary. This boundary level is hard
-    coded as a function of surface pressure and not customizable.
-
-    Args:
-        q: Specific humidity at pressure levels
-        u: Eastward horizontal wind speed at pressure levels
-        v: Northward horizontal wind speed at pressure levels
-        ps: Air pressure at the surface
-        qs: Specific humidity at the surface
-        us: Eastward horizontal wind speed at the surface
-        vs: Northward horizontal wind speed at the surface
-
-    Returns:
-        p interpolated to new pressure levels,
-        q interpolated to new pressure levels,
-        u interpolated to new pressure levels,
-        v interpolated to new pressure levels,
-        Pressure at the division between the two layers
-    """
-    level_ax: int = u.get_axis_num("level")  # type: ignore
-
-    p = u["level"].broadcast_like(u) * 100  # Pa
-    p.attrs["units"] = "Pa"
-
-    _p = p.to_numpy()
-    _u = u.to_numpy()
-    _v = v.to_numpy()
-    _q = q.to_numpy()
-
-    # Insert top of atmosphere values
-    # Assume wind at top same as values at lowest pressure, humidity at top = 0
-    p0 = _p.argmin(axis=level_ax).flatten()[
-        0
-    ]  # index along level axis of minimum pressure
-    _p = np.insert(_p, 0, 0.0, axis=level_ax)
-    _q = np.insert(_q, 0, 0.0, axis=level_ax)
-    _v = np.insert(_v, 0, np.take(_v, p0, axis=level_ax), axis=level_ax)
-    _u = np.insert(_u, 0, np.take(_u, p0, axis=level_ax), axis=level_ax)
-
-    # Insert surface level values
-    _p = np.insert(_p, 0, ps, axis=level_ax)
-    _q = np.insert(_q, 0, qs, axis=level_ax)
-    _v = np.insert(_v, 0, vs, axis=level_ax)
-    _u = np.insert(_u, 0, us, axis=level_ax)
-
-    # Sort arrays by pressure (ascending)
-    _u = sortby_ndarray(_u, _p, axis=level_ax)
-    _v = sortby_ndarray(_v, _p, axis=level_ax)
-    _q = sortby_ndarray(_q, _p, axis=level_ax)
-    _p = sortby_ndarray(_p, _p, axis=level_ax)
-
-    # Insert boundary level values (at a dummy pressure value)
-    p_boundary = 0.72878581 * np.array(ps) + 7438.803223
-    _u = np.insert(_u, 0, interpolate(p_boundary, _p, _u, axis=level_ax), axis=level_ax)
-    _v = np.insert(_v, 0, interpolate(p_boundary, _p, _v, axis=level_ax), axis=level_ax)
-    _q = np.insert(_q, 0, interpolate(p_boundary, _p, _q, axis=level_ax), axis=level_ax)
-    _p = np.insert(_p, 0, p_boundary, axis=level_ax)
-
-    # Sort arrays by pressure once more (ascending)
-    _u = sortby_ndarray(_u, _p, axis=level_ax)
-    _v = sortby_ndarray(_v, _p, axis=level_ax)
-    _q = sortby_ndarray(_q, _p, axis=level_ax)
-    _p = sortby_ndarray(_p, _p, axis=level_ax)
-
-    # Reset level coordinate as its values have become meaningless
-    nlev = np.size(_u, axis=level_ax)
-    levs = np.arange(nlev)
-
-    # reconstruct dataarrays
-    coords = dict(u.coords)
-    coords["level"] = levs
-    u = xr.DataArray(data=_u, dims=u.dims, coords=coords, attrs=u.attrs)
-    v = xr.DataArray(data=_v, dims=v.dims, coords=coords, attrs=v.attrs)
-    q = xr.DataArray(data=_q, dims=q.dims, coords=coords, attrs=q.attrs)
-    p = xr.DataArray(data=_p, dims=p.dims, coords=coords, attrs=p.attrs)
-    pb = xr.DataArray(data=p_boundary, dims=ps.dims, coords=ps.coords)
-
-    # Give the level dim new indexes (just ordered numbers)
-    nlev = u["level"].size
-    u = u.assign_coords(level=np.arange(nlev))
-    v = v.assign_coords(level=np.arange(nlev))
-    q = q.assign_coords(level=np.arange(nlev))
-    p = p.assign_coords(level=np.arange(nlev))
-
-    return q, u, v, p, pb
-
-
-def interp_dp_midpoints(
-    q: xr.DataArray,
-    u: xr.DataArray,
-    v: xr.DataArray,
-    p: xr.DataArray,
-    ps: xr.DataArray,
-):
-    """Interpolate the data to midpoints to allow for integration to two layers.
-
-    Args:
-        q: Specific humidity at levels
-        u: Eastward horizontal wind speed at levels
-        v: Northward horizontal wind speed at levels
-        p: Air pressure at levels
-        ps: Air pressure at the surface
-
-    Returns:
-        Pressure difference between the pressure level bounds,
-        p interpolated to midpoints of pressure levels,
-        q interpolated to midpoints of pressure levels,
-        u interpolated to midpoints of pressure levels,
-        v interpolated to midpoints of pressure levels,
-    """
-
-    # Calculate pressure jump
-    dp = p.diff("level")
-    assert np.all(dp >= 0), "Pressure levels should increase monotonically"
-
-    # Interpolate to midpoints
-    midpoints = 0.5 * (u["level"].to_numpy()[1:] + u["level"].to_numpy()[:-1])
-    dp = dp.assign_coords(level=midpoints)
-    u = u.interp(level=midpoints)
-    v = v.interp(level=midpoints)
-    q = q.interp(level=midpoints)
-    p = p.interp(level=midpoints)
-
-    # mask values below surface
-    above_surface = p < np.array(ps)[None, :, :]
-    u = u.where(above_surface)
-    v = v.where(above_surface)
-    q = q.where(above_surface)
-    p = p.where(above_surface)
-
-    return dp, q, u, v, p
-
-
-def get_input_dates(config):
-    """Dates for pre-processing."""
-    return pd.date_range(
-        start=config.preprocess_start_date,
-        end=config.preprocess_end_date,
-        freq=config.input_frequency,
-    )
-
-
-def prep_experiment(config_file):
-    """Pre-process all data for a given config file.
-
-    This function expects the following configuration settings:
-
-    - preprocess_start_date: formatted as YYYYMMDD, e.g. '20210701'
-    - preprocess_end_date: formatted as YYYYMMDD, e.g. '20210716'
-    - level_type: either "pressure_levels" or "model_levels"
-    - levels: "all" or a list of integers with the desired (model or pressure)
-      levels.
-    - input_folder: path where raw era5 input data can be found, e.g.
-      /home/peter/WAM2layers/era5_2021
-    - preprocessed_data_folder: path where preprocessed data should be stored.
-      This directory will be created if it does not exist. E.g.
-      /home/peter/WAM2layers/preprocessed_data_2021
-    - filename_prefix: Fixed part of filename. This function will infer the
-      variable name and add _ml for model level data. E.g. with prefix =
-      "FloodCase_202107" this function will be able to find
-      FloodCase_202107_ml_u.nc or FloodCase_202107_u.nc and
-      FloodCase_202107_sp.nc
-    """
-    config = Config.from_yaml(config_file)
-
-    for datetime in get_input_dates(config):
-        is_new_day = datetime == datetime.floor("1d")
-        logger.info(datetime)
-
-        # 4d fields
-        levels = config.levels
-        q = load_data("q", datetime, config)  # in kg kg-1
-        u = load_data("u", datetime, config)  # in m/s
-        v = load_data("v", datetime, config)  # in m/s
-        sp = load_data("sp", datetime, config)  # in Pa
-
-        if config.level_type == "model_levels":
-            dp = get_dp_modellevels(sp, levels)
-
-        if config.level_type == "pressure_levels":
-            d2m = load_data("d2m", datetime, config)  # Dew point in K
-            q2m = calculate_humidity(d2m, sp)  # kg kg-1
-            u10 = load_data("u10", datetime, config)  # in m/s
-            v10 = load_data("v10", datetime, config)  # in m/s
-
-            q, u, v, p, pb = extend_pressurelevels(q, u, v, sp, q2m, u10, v10)
-            dp, q, u, v, p = interp_dp_midpoints(q, u, v, p, sp)
-
-        # Calculate column water vapour
-        g = 9.80665  # gravitational accelleration [m/s2]
-        cwv = q * dp / g  # (kg/m2)
-
-        try:
-            # Calculate column water instead of column water vapour
-            tcw = load_data("tcw", datetime, config)  # kg/m2
-            correction = tcw / cwv.sum(dim="level")
-            cw = correction * cwv  # column water (kg/m2)
-            if is_new_day:
-                logger.info(
-                    "Total column water correction:\n"
-                    "    ratio total column water / computed integrated water vapour\n"
-                    f"    mean over grid for this timestep {correction.mean().item():.4f}"
-                )
-
-        except FileNotFoundError:
-            # Fluxes will be calculated based on the column water vapour
-            cw = cwv
-            log_once(
-                logger, "Total column water not available; using water vapour only"
-            )
-
-        # Integrate fluxes and states to upper and lower layer
-        if config.level_type == "model_levels":
-            # TODO: Check if this is a reasonable choice for boundary
-            boundary = 111
-            lower_layer = dp["level"] > boundary
-            upper_layer = ~lower_layer
-
-        if config.level_type == "pressure_levels":
-            upper_layer = p < pb.broadcast_like(p)
-            lower_layer = pb.broadcast_like(p) < p
-
-        # Vertically integrate state over two layers
-        s_lower = cw.where(lower_layer).sum(dim="level")
-        s_upper = cw.where(upper_layer).sum(dim="level")
-
-        # Determine the fluxes
-        fx = u * cw  # eastward atmospheric moisture flux (kg m-1 s-1)
-        fy = v * cw  # northward atmospheric moisture flux (kg m-1 s-1)
-
-        # Vertically integrate fluxes over two layers
-        fx_lower = fx.where(lower_layer).sum(dim="level")  # kg m-1 s-1
-        fy_lower = fy.where(lower_layer).sum(dim="level")  # kg m-1 s-1
-        fx_upper = fx.where(upper_layer).sum(dim="level")  # kg m-1 s-1
-        fy_upper = fy.where(upper_layer).sum(dim="level")  # kg m-1 s-1
-
-        precip, evap = preprocess_precip_and_evap(datetime, config)
-
-        # Combine everything into one dataset
-        ds = (
-            xr.Dataset(
-                {
-                    "fx_upper": fx_upper.assign_attrs(units="kg m-1 s-1"),
-                    "fy_upper": fy_upper.assign_attrs(units="kg m-1 s-1"),
-                    "fx_lower": fx_lower.assign_attrs(units="kg m-1 s-1"),
-                    "fy_lower": fy_lower.assign_attrs(units="kg m-1 s-1"),
-                    "s_upper": s_upper.assign_attrs(units="kg m-2"),
-                    "s_lower": s_lower.assign_attrs(units="kg m-2"),
-                    "evap": evap,
-                    "precip": precip,
-                }
-            )
-            .expand_dims("time")
-            .astype("float32")
-        )
-
-        # Verify that the data meets all the requirements for the model
-        # check_input(ds)
-
-        add_bounds(ds)
-
-        # Add attributes
-        for var in ERA5_INPUT_DATA_ATTRIBUTES:
-            ds[var].attrs.update(ERA5_INPUT_DATA_ATTRIBUTES[var])
-        ds.attrs.update(PREPROCESS_ATTRS)
-
-        # Save preprocessed data
-        filename = f"{datetime.strftime('%Y-%m-%d')}_fluxes_storages.nc"
-        output_path = config.preprocessed_data_folder / filename
-
-        if is_new_day:
-            comp = dict(zlib=True, complevel=8)
-            encoding = {var: comp for var in ds.data_vars}
-            time_encoding = {"units": "seconds since 1900-01-01"}
-            encoding["time"] = time_encoding
-            ds.to_netcdf(
-                output_path, unlimited_dims=["time"], mode="w", encoding=encoding
-            )
-
-        else:
-            append_to_netcdf(output_path, ds, expanding_dim="time")
